+++ src/main/java/com/picpay/transfers/domain/entities/Transfer.java
@@ -4,10 +4,10 @@ import com.picpay.transfers.domain.value_objects.*;
 
 public final class Transfer {
     private Money value;
-    private int payerId;
-    private int payeeId;
+    private long payerId;
+    private long payeeId;
 
-    public Transfer(Money value, int payerId, int payeeId) {
+    public Transfer(Money value, long payerId, long payeeId) {
         this.value = value;
         this.payerId = payerId;
         this.payeeId = payeeId;
